@@ -12,10 +12,10 @@ import subprocess
 import sys
 import typing
 import zipfile
-from collections.abc import Coroutine
+from collections.abc import Awaitable, Callable, Iterable, Sequence
 from math import ceil
 from pathlib import Path
-from typing import Any, Iterable
+from typing import Any
 
 import aiofiles
 import psutil
@@ -24,13 +24,33 @@ import yaml
 from flet import Text
 from lxml import etree, objectify
 
-from commod.helpers.parse_ops import beautify_machina_xml, xml_to_objfy
 from commod.game.data import ENCODING
+from commod.helpers.parse_ops import beautify_machina_xml, xml_to_objfy
 
 logger = logging.getLogger("dem")
 
 SUPPORTED_IMG_TYPES = (".svg", ".png", ".jpg", ".jpeg", ".gif", ".webp", ".bmp")
 RESOLUTION_OPTION_LIST_SIZE = 5
+
+def process_xml_tree(objectify_tree: objectify.ObjectifiedElement,
+                     machina_beautify: bool = True,
+                     use_utf: bool = False) -> bytes:
+    if use_utf:
+        doctype = ""
+        encoding = "utf-8"
+    else:
+        doctype = f'<?xml version="1.0" encoding="{ENCODING}" standalone="yes" ?>'
+        encoding = ENCODING
+
+    etree.indent(objectify_tree, space="    ")
+    xml_string = etree.tostring(
+        objectify_tree,
+        pretty_print=True,
+        xml_declaration=False,
+        doctype=doctype,
+        encoding=encoding)
+
+    return beautify_machina_xml(xml_string) if machina_beautify else xml_string
 
 def write_xml_to_file(
     objectify_tree: objectify.ObjectifiedElement,
@@ -41,32 +61,16 @@ def write_xml_to_file(
     Will format and beautify file in the style very similar to original EM dynamicscene.xml
     files by default. Can skip beautifier and save raw lxml formated file.
     """
-    if use_utf:
-        doctype = ""
-        encoding = "utf-8"
-    else:
-        doctype = f'<?xml version="1.0" encoding="{ENCODING}" standalone="yes" ?>'
-        encoding = ENCODING
-
-    xml_string = etree.tostring(
-        objectify_tree,
-        pretty_print=True,
-        xml_declaration=False,
-        doctype=doctype,
-        # standalone=True,
-        encoding=encoding)
-
-    # TODO: replace("&quot;", "''") maybe?
-
     with Path(path).open("wb") as fh:
-        if machina_beautify:
-            fh.write(beautify_machina_xml(xml_string))
-        else:
-            fh.write(xml_string)
+        fh.write(process_xml_tree(
+            objectify_tree,
+            machina_beautify=machina_beautify,
+            use_utf=use_utf))
 
 
 async def write_xml_to_file_async(
-        objectify_tree: objectify.ObjectifiedElement, path: str,
+        objectify_tree: objectify.ObjectifiedElement,
+        path: str | Path,
         machina_beautify: bool = True,
         use_utf: bool = False) -> None:
     """Asynchronously write ObjectifiedElement tree to file at path.
@@ -75,24 +79,11 @@ async def write_xml_to_file_async(
     dynamicscene.xml files by default. Can skip beautifier and save raw lxml
     formated file.
     """
-    if use_utf:
-        doctype = ""
-        encoding = "utf-8"
-    else:
-        doctype = f'<?xml version="1.0" encoding="{ENCODING}" standalone="yes" ?>'
-        encoding = ENCODING
-
-    xml_string = etree.tostring(
-        objectify_tree,
-        pretty_print=True,
-        doctype=doctype,
-        encoding=encoding)
-
-    async with aiofiles.open(path, mode="wb", encoding=encoding) as fh:
-        if machina_beautify:
-            await fh.write(beautify_machina_xml(xml_string))
-        else:
-            await fh.write(xml_string)
+    async with aiofiles.open(path, mode="wb") as fh:
+        await fh.write(process_xml_tree(
+            objectify_tree,
+            machina_beautify=machina_beautify,
+            use_utf=use_utf))
 
 def open_dir_in_os(directory_path: str | Path) -> None:
     """Open directory in Windows Explorer or OS specific equiavalents."""
@@ -111,14 +102,23 @@ def open_file_in_editor(file_path: str | Path, editor: str, line: int = 1) -> No
     if file_path.exists():
         final_cmd = [editor]
         try:
-            if line != 1 and (editor == "code" or "code.exe" in editor.lower()):
-                final_cmd.extend(["-g", f"{file_path}:{line}"])
-            elif line != 1 and "notepad++" in editor.lower():
-                final_cmd.extend([str(file_path), f"-n{line}"])
-            else:
+            if editor == "code" or "code.exe" in editor.lower():
+                if line != 1:
+                    final_cmd.extend(["-g", f"{file_path}:{line}"])
+                else:
+                    final_cmd.append(str(file_path))
+            elif "notepad++" in editor.lower():
+                if line != 1:
+                    final_cmd.extend([str(file_path), f"-n{line}"])
+                else:
+                    final_cmd.append(str(file_path))
+            elif Path(editor).exists():
                 final_cmd.append(str(file_path))
+            else:
+                return
 
-            subprocess.run(final_cmd, shell=True)  # noqa: S603
+            # looks bad but isn't, we validate both parts of shell invocation to disallow arbitrary execution
+            subprocess.run(final_cmd, shell=True, check=False)  # noqa: S602
         except Exception:
             logger.exception("Unable to open file in editor")
 
@@ -127,13 +127,12 @@ def count_files(directory: str) -> int:
 
     if os.path.isdir(directory):
         for _, _, filenames in os.walk(directory):
-            files.extend(filenames)
+            files.extend(filename for filename in filenames if not filename.startswith("_"))
 
     return len(files)
 
-
 async def copy_from_to_async(from_path_list: list[str],
-                             to_path: str, callback_progbar: callable) -> None:
+                             to_path: str, callback_progbar: Callable) -> None:
     files_count = 0
     for from_path in from_path_list:
         logger.debug(f"Copying files from '{from_path}' to '{to_path}'")
@@ -152,26 +151,68 @@ async def copy_from_to_async(from_path_list: list[str],
                 await callback_progbar(file_num, files_count, sfile, file_size)
                 file_num += 1
 
+async def copy_relative_paths_and_call_async(
+        relative_path: Path,
+        file_num: list[int],
+        base_from_path: str | Path, base_to_path: str | Path,
+        files_count: int,
+        callback_progbar: Callable[[int, int, str, float], Awaitable[None]]) -> None:
+    from_path = Path(base_from_path, relative_path)
+    to_path = Path(base_to_path, relative_path)
+    # TODO: rethink if something this ugly is really required \/
+    # Note: file num is a dirty hack to pass pointer to mutable int value (index of current file)
+    file_size = round(from_path.stat().st_size / 1024, 2)
+    try:
+        await asyncio.to_thread(shutil.copy2, from_path, to_path)
+    except PermissionError:
+        msg = f"Can't overwrite path '{to_path}', this file is blocked by something, possibly opened"
+        raise PermissionError(msg)  # noqa: B904
+    await callback_progbar(file_num[0], files_count, from_path.name, file_size)
+    await asyncio.sleep(0)
+    file_num[0] += 1
 
-async def copy_file_and_call_async(path: str, file_num: list[int],
-                                   single_file: str,
-                                   from_path: str, to_path: str,
-                                   files_count: int,
-                                   callback_progbar: Coroutine) -> None:
+async def copy_file_and_call_async(
+        path: str, file_num: list[int],
+        single_file: str,
+        from_path: str, to_path: str,
+        files_count: int,
+        callback_progbar: Callable[[int, int, str, float], Awaitable[None]]) -> None:
     # TODO: rethink if something this ugly is really required \/
     # Note: file num is a dirty hack to pass pointer to mutable int value (index of current file)
     dest_file = os.path.join(path.replace(from_path, to_path), single_file)
     file_size = round(Path(os.path.join(path, single_file)).stat().st_size / 1024, 2)
     await asyncio.to_thread(shutil.copy2, os.path.join(path, single_file), dest_file)
-    # TODO: describe interface for this type of callback
     await callback_progbar(file_num[0], files_count, single_file, file_size)
     await asyncio.sleep(0)
     file_num[0] += 1
 
 
-async def copy_from_to_async_fast(from_path_list: list[str | Path],
-                                  to_path: str | Path,
-                                  callback_progbar: callable) -> None:
+async def copy_targets_from_to_async(
+        targets_list: Sequence[Path],
+        from_base_path: str | Path,
+        to_base_path: str | Path,
+        callback_progbar: Callable[[int, int, str, float], Awaitable[None]]) -> None:
+
+    from_base_path = Path(from_base_path)
+    to_base_path = Path(to_base_path)
+    files_count = len(targets_list)
+    file_num = []
+    file_num.append(1)
+    for target in targets_list:
+        file_parent = (to_base_path / target).parent
+        file_parent.mkdir(parents=True, exist_ok=True)
+
+    await asyncio.gather(*[
+        copy_relative_paths_and_call_async(
+            target, file_num,
+            from_base_path, to_base_path,
+            files_count,
+            callback_progbar) for target in targets_list])
+
+async def copy_from_to_async_fast(
+        from_path_list: Sequence[str | Path],
+        to_path: str | Path,
+        callback_progbar: Callable[[int, int, str, float], Awaitable[None]]) -> None:
     from_path_list = [str(path_entry) for path_entry in from_path_list]
     to_path = str(to_path)
 
@@ -190,14 +231,15 @@ async def copy_from_to_async_fast(from_path_list: list[str | Path],
             await asyncio.gather(*[
                 copy_file_and_call_async(path, file_num, single_file,
                                          from_path, to_path, files_count,
-                                         callback_progbar) for single_file in filenames])
+                                         callback_progbar) for single_file in filenames
+                                         if not single_file.startswith("_")])
 
 
 async def extract_files_from_zip(
         archive: zipfile.ZipFile,
         file_names: list[str],
         path: str | Path,
-        callback: Coroutine | None = None,
+        callback: Callable | None = None,
         files_num: int = 1) -> None:
     for file_name_raw in file_names:
         file_name = file_name_raw
@@ -209,9 +251,12 @@ async def extract_files_from_zip(
         except UnicodeEncodeError:
             pass
         filepath = Path(path, file_name)
+        if not filepath.parent.is_dir():
+            os.makedirs(filepath.parent, exist_ok=True)
+
         async with aiofiles.open(str(filepath), "wb") as fd:
             await fd.write(data)
-        if callable is not None:
+        if callback is not None:
             await callback(files_num)
 
 
@@ -219,16 +264,16 @@ async def extract_files_from_7z(
         archive: py7zr.SevenZipFile,
         file_names: list[str],
         path: str | Path,
-        callback: Coroutine | None = None,
+        callback: Callable | None = None,
         files_num: int = 1, chunksize: int = 1) -> None:
     archive.reset()
     archive.extract(path, targets=file_names)
-    if callable is not None:
+    if callback is not None:
         await callback(files_num, chunksize)
         await asyncio.sleep(0)
 
 
-async def extract_archive_from_to(archive_path: str, to_path: str, callback: Coroutine | None = None,
+async def extract_archive_from_to(archive_path: str, to_path: str, callback: Callable | None = None,
                           loading_text: Text | None = None) -> None:
     extension = Path(archive_path).suffix
     match extension:
@@ -241,7 +286,7 @@ async def extract_archive_from_to(archive_path: str, to_path: str, callback: Cor
 
 
 async def extract_zip_from_to(archive_path: str | Path, to_path: str | Path,
-                              callback: Coroutine | None = None,
+                              callback: Callable | None = None,
                               loading_text: Text | None = None) -> None:
     os.makedirs(to_path, exist_ok=True)
     with zipfile.ZipFile(archive_path, "r") as archive:
@@ -257,6 +302,7 @@ async def extract_zip_from_to(archive_path: str | Path, to_path: str | Path,
         if chunksize == 0:
             chunksize = 1
         tasks = []
+
         for file in archive.filelist:
             file_path = file.filename
             if file.is_dir():
@@ -298,7 +344,7 @@ async def extract_zip_from_to(archive_path: str | Path, to_path: str | Path,
 
 
 async def extract_7z_from_to(archive_path: str | Path, to_path: str | Path,
-                             callback: Coroutine | None = None,
+                             callback: Callable | None = None,
                              loading_text: Text | None = None) -> None:
     os.makedirs(to_path, exist_ok=True)
     with py7zr.SevenZipFile(str(archive_path), "r") as archive:
@@ -306,7 +352,7 @@ async def extract_7z_from_to(archive_path: str | Path, to_path: str | Path,
             info = archive.archiveinfo()
             loading_text.value = (f"[{info.method_names[0]}] "
                                   f"{info.size/1024/1024:.1f}MB -> "
-                                  f"{info.uncompressed/1024/1024:.1f}MB")
+                                  f"{info.uncompressed/1024/1024:.1f}MB") # type: ignore (is actually int)
             loading_text.update()
             await asyncio.sleep(0)
         all_files = archive.files
@@ -324,13 +370,13 @@ async def extract_7z_from_to(archive_path: str | Path, to_path: str | Path,
         archive_size = archive.archiveinfo().uncompressed
         # chunk extraction for every 32MB of internal data to show some kind of progress
         # if file is big, extract it in 5 chunks
-        chunk_file_size = archive_size / 5
+        chunk_file_size = archive_size / 5 # type: ignore (is actually int)
         default_chunk_file_size = 1024 * 1024 * 32
 
         if chunk_file_size > default_chunk_file_size:
             workers = round(archive_size / chunk_file_size)
         else:
-            workers = round(archive_size / default_chunk_file_size)
+            workers = round(archive_size / default_chunk_file_size) # type: ignore (is actually int)
 
         if workers == 0:
             workers = 1
@@ -390,15 +436,16 @@ def get_internal_file_path(file_name: str | Path) -> Path:
 def patch_offsets(f: typing.BinaryIO,
                   offsets_dict: dict, enlarge_coeff: float = 1.0,
                   raw_strings: bool = False) -> None:
-    for offset in offsets_dict:
-        new_value = offsets_dict[offset]
+    for offset, new_value in offsets_dict.items():
         f.seek(offset)
 
         # type equality used instead of isinstance because isinstance(True, int) is True, it's an error here
         if type(new_value) == int:
             if not math.isclose(enlarge_coeff, 1.0):
-                new_value = round(new_value * enlarge_coeff)
-            f.write(struct.pack("i", new_value))
+                final_value = round(new_value * enlarge_coeff)
+            else:
+                final_value = new_value
+            f.write(struct.pack("i", final_value))
         elif type(new_value) == str:
             if raw_strings:  # write as is, binary insert strings
                 f.write(bytes.fromhex(new_value))
@@ -406,8 +453,10 @@ def patch_offsets(f: typing.BinaryIO,
                 f.write(struct.pack("<L", int(new_value, base=16)))
         elif type(new_value) == float:
             if not math.isclose(enlarge_coeff, 1.0):
-                new_value = round(new_value * enlarge_coeff)
-            f.write(struct.pack("f", new_value))
+                final_value = round(new_value * enlarge_coeff)
+            else:
+                final_value = new_value
+            f.write(struct.pack("f", final_value))
         elif type(new_value) == bool:
             f.write(struct.pack("b", new_value))
         elif type(new_value) == tuple:
@@ -416,8 +465,8 @@ def patch_offsets(f: typing.BinaryIO,
             raise TypeError("Unsuported type given")
 
 
-def get_config(root_dir: str) -> objectify.ObjectifiedElement:
-    return xml_to_objfy(os.path.join(root_dir, "data", "config.cfg"))
+def get_config(root_dir: str | Path) -> objectify.ObjectifiedElement:
+    return xml_to_objfy(Path(root_dir, "data", "config.cfg"))
 
 
 def running_in_venv() -> bool:
