@@ -4,20 +4,23 @@ import tempfile
 from pathlib import Path
 
 import flet as ft
-from flet import IconButton, Image, Page, Theme, VisualDensity
 
 from commod.game.data import get_title
 from commod.game.environment import GameCopy, InstallationContext
 from commod.gui import common_widgets as cw
-from commod.gui.app_widgets import App
+from commod.gui.app_widgets import App, PageHelper
 from commod.gui.config import AppSections, Config
 from commod.helpers.file_ops import get_internal_file_path
 from commod.helpers.parse_ops import init_input_parser
 from commod.localisation.service import tr
 
 
-async def main(page: Page) -> None:
+async def main(page: ft.Page) -> None:
     options = init_input_parser().parse_args()
+
+    page.fonts = {
+        "Fira Code": str(get_internal_file_path("assets/fonts/FiraCode-Regular.ttf"))
+    }
 
     page.window.title_bar_hidden = True
     page.title = "ComMod"
@@ -27,8 +30,8 @@ async def main(page: Page) -> None:
     page.padding = 0
 
     page.theme_mode = ft.ThemeMode.SYSTEM
-    page.theme = Theme(color_scheme_seed="#FFA500", visual_density=VisualDensity.COMPACT)
-    page.dark_theme = Theme(color_scheme_seed="#FFA500", visual_density=VisualDensity.COMPACT)
+    page.theme = ft.Theme(color_scheme_seed="#FFA500", visual_density=ft.VisualDensity.COMPACT)
+    page.dark_theme = ft.Theme(color_scheme_seed="#FFA500", visual_density=ft.VisualDensity.COMPACT)
 
     conf = Config(page)
     # at the end of execution, commod tries to create config near itself
@@ -57,6 +60,7 @@ async def main(page: Page) -> None:
     else:
         distribution_dir = conf.current_distro
 
+    install_context: InstallationContext | None = None
     if distribution_dir:
         try:
             install_context = InstallationContext(
@@ -66,7 +70,8 @@ async def main(page: Page) -> None:
             install_context.setup_loggers()
         except Exception:
             # TODO: handle individuals exceptions properly if they are not caught lower
-            install_context.logger.exception("[Distro loading error]")
+            if install_context is not None:
+                install_context.logger.exception("[Distro loading error]")
 
             install_context = InstallationContext(dev_mode=options.dev)
             install_context.setup_loggers(stream_only=True)
@@ -83,7 +88,8 @@ async def main(page: Page) -> None:
         # if app.config.known_games:
         target_dir = conf.current_game
         # else:
-        # TODO: rework for this default to work as expected, should detect the game and add to the list as current
+        # TODO: rework this default to work as expected:
+        # should detect the game and add to the list as current
         # target_dir = InstallationContext.get_local_config_path()
 
 
@@ -100,11 +106,14 @@ async def main(page: Page) -> None:
     else:
         game = GameCopy()
 
+    page_helper = PageHelper()
+
     # TODO: pass 'dev' options further, it's needed in case of changing the context
     app = App(context=install_context,
               game=game,
               config=conf,
-              page=page)
+              page=page,
+              page_helper=page_helper)
 
     page.window.on_event = app.wrapped_on_window_event
     page.theme_mode = conf.init_theme
@@ -118,7 +127,6 @@ async def main(page: Page) -> None:
 
     app.create_sections()
 
-    page.theme_icon_btn = ft.Ref[IconButton]()
     theme_icon = ft.Icons.BRIGHTNESS_AUTO
     match page.theme_mode:
         case ft.ThemeMode.SYSTEM:
@@ -150,20 +158,21 @@ async def main(page: Page) -> None:
                 label=tr("download").capitalize()
             ),
             ft.NavigationRailDestination(
-                icon=ft.Icons.CONSTRUCTION_OUTLINED,
-                selected_icon=ft.Icons.CONSTRUCTION,
-                label=tr("modding_tools").capitalize(),
-            ),
-            ft.NavigationRailDestination(
                 icon=ft.Icons.SETTINGS_OUTLINED,
                 selected_icon=ft.Icons.SETTINGS,
                 label=tr("settings").capitalize()
-            )
+            ),
+            ft.NavigationRailDestination(
+                icon=ft.Icons.CONSTRUCTION_OUTLINED,
+                selected_icon=ft.Icons.CONSTRUCTION,
+                label=tr("modding_tools").capitalize(),
+                visible=app.config.modder_mode
+            ),
         ],
         trailing=ft.IconButton(
             icon=theme_icon,
             on_click=app.change_theme_mode,
-            ref=page.theme_icon_btn,
+            ref=page_helper.theme_icon_btn,
             selected_icon_color=ft.Colors.ON_SURFACE_VARIANT,
             tooltip=ft.Tooltip(
                 message=tr("theme_mode"),
@@ -172,31 +181,29 @@ async def main(page: Page) -> None:
     )
     app.rail = rail
 
-    page.minimize_btn = ft.Ref[IconButton]()
-    page.maximize_btn = ft.Ref[IconButton]()
-    # title bar to replace system one
     page.remove(loading_ring)
+
     page.add(
         ft.Row(
             [ft.WindowDragArea(ft.Container(
                  ft.Row([
-                     Image(src=get_internal_file_path("assets/icons/dem_logo.svg"),
+                     ft.Image(src=str(get_internal_file_path("assets/icons/dem_logo.svg")),
                            width=20,
                            height=20,
                            fit=ft.ImageFit.COVER),
                      ft.Text(get_title(), size=13, weight=ft.FontWeight.W_500),
                      ft.Text("[dev]", size=13, color=ft.Colors.ERROR, weight=ft.FontWeight.BOLD,
-                             visible=install_context.dev_mode or app.config.modder_mode),
+                             visible=install_context.dev_mode),
                      ]), padding=6),
                      expand=True),
              cw.TitleButton(ft.Icons.MINIMIZE_ROUNDED,
                             on_click=app.minimize,
                             icon_size=20,
-                            ref=page.minimize_btn),
+                            ref=page_helper.minimize_btn),
              cw.TitleButton(ft.Icons.CHECK_BOX_OUTLINE_BLANK_ROUNDED,
                             on_click=app.maximize,
                             icon_size=17,
-                            ref=page.maximize_btn),
+                            ref=page_helper.maximize_btn),
              cw.TitleButton(ft.Icons.CLOSE_ROUNDED,
                             on_click=app.finalize,
                             icon_size=22,
@@ -207,14 +214,14 @@ async def main(page: Page) -> None:
         )
     )
 
-    app.content_column = ft.Container(loading_ring,
-                                      expand=True,
-                                      alignment=ft.alignment.top_center,
-                                      margin=ft.margin.only(left=0, right=0))
+    app.content_container.content = loading_ring
+    app.content_container.expand = True
+    app.content_container.alignment = ft.alignment.top_center
+    app.content_container.margin = ft.margin.only(left=0, right=0)
 
     # add application's root control to the page
     page.add(
-        ft.Container(ft.Row([rail, app.content_column]),
+        ft.Container(ft.Row([rail, app.content_container]),
                      expand=True,
                      padding=ft.padding.only(left=10, right=10, bottom=10)
                      )
@@ -224,12 +231,14 @@ async def main(page: Page) -> None:
 
     if app.context.under_windows:
         app.context.current_session.load_steam_game_paths()
+        if app.context.current_session.steam_parsing_error:
+            app.logger.debug(app.context.current_session.steam_parsing_error)
 
     if need_quick_start:
         app.logger.debug("showing quick start")
         # modern settings screen has a built-in flow for quick start
-        app.content_column.content = None
-        app.content_column.update()
+        app.content_container.content = None
+        app.content_container.update()
         await app.show_settings()
     else:
         await app.load_distro_async()
@@ -247,19 +256,18 @@ async def main(page: Page) -> None:
             app.config.add_game_to_config(app.game.game_root_path, Path(app.game.game_root_path).stem)
             app.logger.debug(f"Added game passed to args to the config: {app.game.game_root_path}")
         # select_game_from_home
-        app.content_column.content = None
-        app.content_column.update()
+        app.content_container.content = None
+        app.content_container.update()
         await app.change_page(index=app.config.current_section)
 
     if "NUITKA_ONEFILE_PARENT" in os.environ:
         splash_filename = os.path.join(
             tempfile.gettempdir(),
-            "onefile_%d_splash_feedback.tmp" % int(os.environ["NUITKA_ONEFILE_PARENT"]),
+            "onefile_%d_splash_feedback.tmp" % int(os.environ["NUITKA_ONEFILE_PARENT"]),  # noqa: UP031
         )
 
         if os.path.exists(splash_filename):
-            os.unlink(splash_filename)
-
+          os.unlink(splash_filename)
 
     page.update()
 
