@@ -1,9 +1,9 @@
 import asyncio
+import contextlib
 import json
 import logging
-from itertools import batched
-from pathlib import Path
 import time
+from pathlib import Path
 
 import flet as ft
 from lxml import etree, objectify
@@ -11,18 +11,127 @@ from lxml import etree, objectify
 from commod.gui import app_widgets
 from commod.gui import common_widgets as cw
 from commod.helpers import file_ops, parse_ops
-from commod.localisation.service import tr
-from commod.tools import xml_diff
-from commod.tools import xml_merge
-from commod.tools.xml_merge import ActionType, Command
+from commod.localisation.service import SupportedLanguages, get_current_lang, tr
+from commod.tools import xml_diff, xml_merge
+from commod.tools.xml_helpers import ActionType, Command, InvalidMergeCommandError
 
 logger = logging.getLogger("dem")
 
-class ModdingTools(ft.Column):
+class ModdingTools(ft.Tabs):
+    def __init__(self, app: "app_widgets.App", **kwargs) -> None:
+        kwargs.setdefault("padding", ft.padding.all(10))
+        kwargs.setdefault("divider_height", 1)
+        kwargs.setdefault("animation_duration", 25)
+        super().__init__(**kwargs)
+
+        self.refreshing = False
+
+            # ft.Text(tr("merge_mod_creation").capitalize(),
+            #         theme_style=ft.TextThemeStyle.TITLE_MEDIUM),
+            # ft.Divider(height=5, color=ft.Colors.TRANSPARENT),
+
+        self.tabs = [
+            ft.Tab(
+                text=tr("merge_mod_creation").capitalize(),
+                content=ft.Container(
+                    content=MergeTool(app), alignment=ft.alignment.center
+                ),
+                height=35
+            ),
+            ft.Tab(
+                text=tr("documentation"),
+                content=ft.Container(
+                    content=MergeModsDocs(app), alignment=ft.alignment.top_left,
+                    padding=ft.padding.all(10)
+                ),
+                height=35
+            ),
+        ]
+
+
+class MergeModsDocs(ft.Row):
     def __init__(self, app: "app_widgets.App", **kwargs) -> None:
         super().__init__(**kwargs)
-        self.refreshing = False
+        self.md_container = ft.Ref[ft.Container]()
+        self.btn_list = ft.Ref[ft.Column]()
+        self.doc_content_container = ft.Column([
+            ft.Container(
+                None,
+                ref=self.md_container,
+                padding=ft.padding.only(left=15, right=10),
+                margin=ft.margin.only(right=10),
+                ),
+            ],
+            alignment=ft.MainAxisAlignment.START,
+            scroll=ft.ScrollMode.ADAPTIVE,
+            expand=True,
+        )
+        self.merge_docs_list = [
+            "whats_merge_mod",
+            "using_differ",
+            "types_of_commands"
+        ]
+
+        self.controls = [
+            ft.Column(controls=[
+                ft.Container(ft.Text(tr("doc_sections"),
+                                     weight=ft.FontWeight.BOLD),
+                                     padding=ft.padding.only(left=10)),
+                ft.Column([
+                    ft.TextButton(
+                        tr(key),
+                        on_click=self.select_document,
+                        style=ft.ButtonStyle(visual_density=ft.VisualDensity.COMPACT),
+                        data=key) for key in self.merge_docs_list
+                    ],
+                    spacing=2,
+                    ref=self.btn_list
+                )
+            ], width=190, expand_loose=True),
+            ft.VerticalDivider(),
+            self.doc_content_container
+        ]
+
+        first_doc = self.btn_list.current.controls[0]
+        if isinstance(first_doc, ft.TextButton) and first_doc.data:
+            first_doc.icon = ft.Icons.SUBDIRECTORY_ARROW_RIGHT
+            self.md_container.current.content = self.get_md_doc(first_doc.data)
+
+    def get_md_doc(self, name: str) -> ft.Markdown:
+        try:
+            lang = get_current_lang()
+            if not file_ops.get_internal_file_path(f"assets/docs/{name}_{lang}.md").exists():
+                lang = SupportedLanguages.ENG.value
+            with open(file_ops.get_internal_file_path(f"assets/docs/{name}_{lang}.md"),
+                      encoding="utf-8") as fh:
+                md1 = fh.read()
+                return cw.MarkdownWithCode(
+                    md1,
+                    selectable=True,
+                    )
+        except FileNotFoundError:
+            return ft.Markdown("NO CONTENT AVAILABLE", expand=True)
+
+    def select_document(self, e: ft.ControlEvent) -> None:
+        if not isinstance(e.control, ft.TextButton):
+            return
+
+        if e.control.data:
+            self.md_container.current.content = self.get_md_doc(e.control.data)
+            self.md_container.current.update()
+            for btn in self.btn_list.current.controls:
+                if isinstance(btn, ft.TextButton):
+                    btn.icon = None
+                    # btn.update()
+            e.control.icon = ft.Icons.SUBDIRECTORY_ARROW_RIGHT
+            self.btn_list.current.update()
+
+class MergeTool(ft.Column):
+    def __init__(self, app: "app_widgets.App", **kwargs) -> None:
+        kwargs.setdefault("spacing", 2)
+        super().__init__(**kwargs)
         self.app = app
+        self.ctrl_pressed = False
 
         diff_guides_path = file_ops.get_internal_file_path("assets/diff_guides.json")
         self.differs = {}
@@ -43,7 +152,10 @@ class ModdingTools(ft.Column):
         self.command_preview = CommandPreview(modding_tools=self)
         self.current_command_card: CommandCard | None = None
         self.commands_view = ft.ListView(
-            controls=[ft.Placeholder(color=ft.Colors.SECONDARY_CONTAINER)], padding=ft.padding.only(right=10))
+            controls=[ft.Container(
+                 ft.Placeholder(color=ft.Colors.SECONDARY_CONTAINER),
+                 padding=ft.padding.symmetric(vertical=3))],
+            padding=ft.padding.only(right=10))
         self.commands_container = ft.Container(
             self.commands_view, expand=True)
 
@@ -71,7 +183,7 @@ class ModdingTools(ft.Column):
         self.output_path_field = cw.TextField(
             label=tr("enter_path_to_output"),
             dense=True,
-            expand=5,
+            expand=6,
             )
 
         self.source_tree: objectify.ObjectifiedElement | None = None
@@ -94,7 +206,6 @@ class ModdingTools(ft.Column):
             tr("save_all"), on_click=self.save_all_ask, disabled=True)
         self.calculate_diff_btn = ft.OutlinedButton(
             tr("calculate_diff"), scale=1.1, on_click=self.calculate_diff)
-
 
     def toggle_preload(self, e: ft.ControlEvent) -> None:
         self.preload_commands_field.disabled = not e.control.value
@@ -124,6 +235,10 @@ class ModdingTools(ft.Column):
                 card.checkbox.value = False
         self.commands_view.update()
 
+    def close_bottom_sheet(self, e: ft.ControlEvent) -> None:
+        with contextlib.suppress(ValueError):
+            self.app.page.overlay.remove(e.control)
+
     def show_bottom_sheet(self, text: str) -> None:
         bs = ft.BottomSheet(
             ft.Container(
@@ -133,6 +248,7 @@ class ModdingTools(ft.Column):
                 padding=20
             ),
             open=True,
+            on_dismiss=self.close_bottom_sheet
         )
         self.app.page.open(bs)
 
@@ -144,16 +260,25 @@ class ModdingTools(ft.Column):
         if Path(self.output_path_field.value).exists():
             await self.app.show_modal(text=tr("overwrite_file_are_you_sure"),
                                       additional_text=self.output_path_field.value,
-                                      on_yes=self.save_selected)
+                                      on_yes=self.save_selected,
+                                      on_no=self.app.close_alert)
         else:
             await self.save_selected()
 
     async def save_selected(self, e: ft.ControlEvent | None = None) -> None:
-        await self.app.close_alert()
-        try:
-            out_path = Path(self.output_path_field.value)
+        self.app.close_alert()
+        if self.source_tree is None:
+            return
 
+        try:
+            if not self.output_path_field.value:
+                return
+
+            out_path = Path(self.output_path_field.value)
             if not out_path.parent.is_dir():
+                await self.app.show_alert(tr("saving_commands_error"),
+                                          f"Directory doesn't exist: {out_path.parent}",
+                                          allow_copy=True)
                 return
         except Exception as ex:
             await self.app.show_alert(tr("saving_commands_error"), str(ex),
@@ -174,9 +299,8 @@ class ModdingTools(ft.Column):
         self.show_bottom_sheet(tr("num_commands_saved", num_cmds=str(len(cmds))))
 
         commands_xml = xml_diff.Differ.serialize_commands(cmds, root_tag=str(self.source_tree.tag))
-        file_ops.write_xml_to_file(commands_xml, out_path,
-                                   machina_beautify=True,
-                                   use_utf=False)
+        await file_ops.write_xml_to_file_async(
+            commands_xml, out_path, machina_beautify=True, use_utf=False)
 
     async def save_all_ask(self, e: ft.ControlEvent) -> None:
         if not self.output_path_field.value or self.source_tree is None:
@@ -186,17 +310,26 @@ class ModdingTools(ft.Column):
         if Path(self.output_path_field.value).exists():
             await self.app.show_modal(text=tr("overwrite_file_are_you_sure"),
                                       additional_text=self.output_path_field.value,
-                                      on_yes=self.save_all)
+                                      on_yes=self.save_all,
+                                      on_no=self.app.close_alert)
         else:
             await self.save_all()
 
 
     async def save_all(self, e: ft.ControlEvent | None = None) -> None:
-        await self.app.close_alert()
-        try:
-            out_path = Path(self.output_path_field.value)
+        self.app.close_alert()
+        if self.source_tree is None:
+            return
 
+        try:
+            if not self.output_path_field.value:
+                return
+
+            out_path = Path(self.output_path_field.value)
             if not out_path.parent.is_dir():
+                await self.app.show_alert(tr("saving_commands_error"),
+                                          f"Directory doesn't exist: {out_path.parent}",
+                                          allow_copy=True)
                 return
         except Exception as ex:
             await self.app.show_alert(tr("saving_commands_error"), str(ex),
@@ -212,78 +345,84 @@ class ModdingTools(ft.Column):
         self.show_bottom_sheet(tr("num_commands_saved", num_cmds=str(len(cmds))))
 
         commands_xml = xml_diff.Differ.serialize_commands(cmds, root_tag=str(self.source_tree.tag))
-        file_ops.write_xml_to_file(commands_xml, out_path, machina_beautify=True, use_utf=False)
+        await file_ops.write_xml_to_file_async(commands_xml, out_path, machina_beautify=True, use_utf=False)
 
     def open_path(self, path_field: ft.TextField) -> None:
         if path_field.value:
+            if not Path(path_field.value).exists():
+                self.app.page.run_task(
+                    self.app.show_alert,
+                    f'{tr("file_doesnt_exist")}:\n`{path_field.value}`')
+                return
             file_ops.open_file_in_editor(path_field.value, editor=self.app.config.code_editor)
 
     def build(self) -> None:
         self.horizontal_alignment=ft.CrossAxisAlignment.CENTER
         self.controls = [
-            ft.Text(tr("merge_mod_creation").capitalize(),
-                 theme_style=ft.TextThemeStyle.TITLE_MEDIUM),
-                ft.Row([self.source_path_field,
-                        ft.IconButton(icon=ft.Icons.EDIT_NOTE,
-                                      on_click=lambda _: self.open_path(path_field=self.source_path_field),
-                                      tooltip=tr("open_in_editor")),
-                        cw.BrowseFileButton(
-                            tr("choose_path"), self.source_path_field,
-                            allowed_extensions=["xml", "ssl"], expand=1,
-                            initial_dir=self.app.config.last_differ_source)], spacing=5),
-                ft.Row([self.modded_path_field,
-                        ft.IconButton(icon=ft.Icons.EDIT_NOTE,
-                                      on_click=lambda _: self.open_path(path_field=self.modded_path_field),
-                                      tooltip=tr("open_in_editor")),
-                        cw.BrowseFileButton(
-                            tr("choose_path"), self.modded_path_field,
-                            allowed_extensions=["xml", "ssl"], expand=1,
-                            initial_dir=self.app.config.last_differ_modded)], spacing=5),
-                ft.Row([self.preload_checkbox,
-                        self.preload_commands_field,
-                        cw.BrowseFileButton(
-                            tr("choose_path"), self.preload_commands_field,
-                            allowed_extensions=["xml"], expand=1,
-                            allow_multiple=True,
-                            initial_dir=self.app.config.last_differ_modded)], spacing=5),
-                ft.Row([
-                    ft.Row([ft.Text(f"{tr("preloaded_commands")}: "), self.preloaded_counter], spacing=0),
-                    self.calculate_diff_btn,
-                    ft.Row([ft.Text(f"{tr("nodes_processed")}: "), self.node_counter], spacing=0),
-                    ft.Row([ft.Text(f"{tr("parsed_commands")}: "), self.command_counter], spacing=0),
-                    ], alignment=ft.MainAxisAlignment.CENTER, spacing=30),
-                ft.Divider(height=0, thickness=2),
-                ft.Row([
-                    ft.Column([
-                        ft.Row([ft.Text(tr("command_list"))],
-                               alignment=ft.MainAxisAlignment.CENTER),
-                        self.commands_container,
-                        ], expand=1, horizontal_alignment=ft.CrossAxisAlignment.CENTER),
-                    ft.Column([
-                        ft.Text(tr("command_preview")),
-                        self.command_preview,
-                        ], expand=1,
-                        horizontal_alignment=ft.CrossAxisAlignment.CENTER),
-                    ft.VerticalDivider(),
-                    ft.Column([
-                        ft.Text(tr("source_node")),
-                        self.source_node_view,
-                        ], expand=1, horizontal_alignment=ft.CrossAxisAlignment.CENTER),
-                    ft.Column([
-                        ft.Text(tr("moddified_node")),
-                        self.modded_node_view,
-                        ], expand=1, horizontal_alignment=ft.CrossAxisAlignment.CENTER),
-                ], expand=True, spacing=5, alignment=ft.MainAxisAlignment.CENTER),
-                ft.Row([self.select_all_btn, self.deselect_all_btn, self.invert_select_btn]),
-                ft.Row([self.output_path_field,
-                        ft.IconButton(icon=ft.Icons.EDIT_NOTE,
-                                      on_click=lambda _: self.open_path(path_field=self.output_path_field),
-                                      tooltip=tr("open_in_editor")),
-                        cw.BrowseFileButton(
-                            tr("choose_path"), self.output_path_field,
-                            mode="save", allowed_extensions=["xml"],
-                            expand=1, initial_dir=self.app.config.last_differ_modded),
-                        self.save_selected_btn, self.save_all_btn]),
+            ft.Divider(height=5, color=ft.Colors.TRANSPARENT),
+            ft.Row([self.source_path_field,
+                    ft.IconButton(icon=ft.Icons.EDIT_NOTE,
+                                  on_click=lambda _: self.open_path(path_field=self.source_path_field),
+                                  tooltip=tr("open_in_editor")),
+                    cw.BrowseFileButton(
+                        tr("choose_path"), self.source_path_field,
+                        allowed_extensions=["xml", "ssl"], expand=1,
+                        initial_dir=self.app.config.last_differ_source)], spacing=5),
+            ft.Row([self.modded_path_field,
+                    ft.IconButton(icon=ft.Icons.EDIT_NOTE,
+                                  on_click=lambda _: self.open_path(path_field=self.modded_path_field),
+                                  tooltip=tr("open_in_editor")),
+                    cw.BrowseFileButton(
+                        tr("choose_path"), self.modded_path_field,
+                        allowed_extensions=["xml", "ssl"], expand=1,
+                        initial_dir=self.app.config.last_differ_modded)], spacing=5),
+            ft.Row([self.preload_checkbox,
+                    self.preload_commands_field,
+                    cw.BrowseFileButton(
+                        tr("choose_path"), self.preload_commands_field,
+                        allowed_extensions=["xml"], expand=1,
+                        allow_multiple=True,
+                        initial_dir=self.app.config.last_differ_modded)], spacing=5),
+            ft.Row([
+                ft.Row([ft.Text(f"{tr("preloaded_commands")}: "), self.preloaded_counter], spacing=0),
+                self.calculate_diff_btn,
+                ft.Row([ft.Text(f"{tr("nodes_processed")}: "), self.node_counter], spacing=0),
+                ft.Row([ft.Text(f"{tr("parsed_commands")}: "), self.command_counter], spacing=0),
+                ], alignment=ft.MainAxisAlignment.CENTER, spacing=30),
+            ft.Divider(height=10, thickness=2),
+            ft.Row([
+                ft.Column([
+                    ft.Row([ft.Text(tr("command_list"))],
+                           alignment=ft.MainAxisAlignment.CENTER),
+                    self.commands_container,
+                    ], expand=1, horizontal_alignment=ft.CrossAxisAlignment.CENTER),
+                ft.Column([
+                    ft.Text(tr("command_preview")),
+                    self.command_preview,
+                    ], expand=1,
+                    horizontal_alignment=ft.CrossAxisAlignment.CENTER),
+                ft.VerticalDivider(),
+                ft.Column([
+                    ft.Text(tr("source_node")),
+                    self.source_node_view,
+                    ], expand=1, horizontal_alignment=ft.CrossAxisAlignment.CENTER),
+                ft.Column([
+                    ft.Text(tr("moddified_node")),
+                    self.modded_node_view,
+                    ], expand=1, horizontal_alignment=ft.CrossAxisAlignment.CENTER),
+            ], expand=True, spacing=5, alignment=ft.MainAxisAlignment.CENTER),
+            ft.Divider(height=5, color=ft.Colors.TRANSPARENT),
+            ft.Row([self.select_all_btn, self.deselect_all_btn, self.invert_select_btn]),
+            ft.Divider(height=5, color=ft.Colors.TRANSPARENT),
+            ft.Row([self.output_path_field,
+                    ft.IconButton(icon=ft.Icons.EDIT_NOTE,
+                                  on_click=lambda _: self.open_path(path_field=self.output_path_field),
+                                  tooltip=tr("open_in_editor")),
+                    cw.BrowseFileButton(
+                        tr("choose_path"), self.output_path_field,
+                        mode="save", allowed_extensions=["xml"],
+                        initial_dir=self.app.config.last_differ_modded),
+                    self.save_selected_btn, self.save_all_btn]),
         ]
 
     def preview_current_node(self) -> None:
@@ -295,7 +434,12 @@ class ModdingTools(ft.Column):
         cmd = self.current_command_card.command
         self.command_preview.update()
 
-        self.source_node_view.node = cmd.source_node if cmd.source_node is not None else None
+        if cmd.source_node is not None:
+            self.source_node_view.node = cmd.source_node
+            self.source_node_view.count = cmd.existing_count
+        else:
+            self.source_node_view.node = None
+
         self.source_node_view.file_path = self.source_path
         self.source_node_view.update()
 
@@ -312,6 +456,10 @@ class ModdingTools(ft.Column):
         self.current_command_card = e.control.parent
         self.preview_current_node()
 
+    def secondary_click_command(self, e: ft.ControlEvent) -> None:
+        e.control.parent.checkbox.value = not e.control.parent.checkbox.value
+        e.control.parent.update()
+
     def cleanup(self, e: ft.ControlEvent | None = None) -> None:
         self.commands_container.content = None
         self.commands_container.update()
@@ -321,8 +469,8 @@ class ModdingTools(ft.Column):
         self.preloaded_commands.clear()
         self.command_preview.update()
 
-        self.source_node_view.content = ft.Placeholder(color=ft.Colors.SECONDARY_CONTAINER)
-        self.modded_node_view.content = ft.Placeholder(color=ft.Colors.SECONDARY_CONTAINER)
+        self.source_node_view.reset()
+        self.modded_node_view.reset()
         self.source_node_view.update()
         self.modded_node_view.update()
 
@@ -343,7 +491,7 @@ class ModdingTools(ft.Column):
         if self.modded_path_field.value and Path(self.modded_path_field.value).exists():
             self.app.config.set_last_differ_modded(Path(self.modded_path_field.value).parent)
 
-    async def calculate_diff(self, e: ft.ControlEvent) -> None:
+    async def calculate_diff(self, e: ft.ControlEvent) -> None:  # noqa: PLR0911
         if not self.source_path_field.value or not self.modded_path_field.value:
             await self.app.show_alert(tr("need_two_paths_for_comparison"))
             return
@@ -380,21 +528,41 @@ class ModdingTools(ft.Column):
                     "Can't produce diff for trees with different root tags: "
                     f"'{self.source_tree.tag}' vs '{self.modded_tree.tag}'")
 
+            if self.source_tree.xpath(".//*[@_Action]"):
+                raise ValueError(
+                    f"Can't produce diff for file containing commands:\n> {self.source_path_field.value}"
+                )
+            if self.modded_tree.xpath(".//*[@_Action]"):
+                raise ValueError(
+                    f"Can't produce diff for file containing commands:\n> {self.modded_path_field.value}"
+                )
+
             if self.preload_commands_field.value and not self.preload_commands_field.disabled:
                 for source_file_path in self.preload_commands_field.value.split(","):
                     source_file = Path(source_file_path.strip())
                     if not source_file.exists() or source_file.suffix != ".xml":
+                        await self.app.show_alert(
+                            f'{tr("command_path_doesnt_exist")}:\n\n`{source_file}`')
                         logger.warning(f"Incorrect file found in preload commands: '{source_file}', skipping")
-                        continue
+                        self.cleanup()
+                        raise InvalidMergeCommandError
                     preloaded_commands = parse_ops.xml_to_objfy(source_file)
                     if preloaded_commands.tag != self.source_tree.tag:
-                        await self.app.show_alert(tr("incorrect_commands_for_source",
-                                                     cmd_path=source_file_path))
+                        await self.app.show_alert(
+                            tr("incorrect_commands_for_source",
+                            cmd_path=source_file_path))
                     else:
-                        commands = xml_merge.parse_command_tree(
-                            preloaded_commands, merge_author=source_file.name)
-                        self.preloaded_commands[str(source_file)] = commands
-
+                        try:
+                            commands = xml_merge.parse_command_tree(
+                                preloaded_commands, merge_author=source_file.name)
+                            self.preloaded_commands[str(source_file)] = commands
+                        except InvalidMergeCommandError as ex:
+                            await self.app.show_alert(tr("existing_command_reading_error") +
+                                                      f":\n\n{source_file_path}\n",
+                                                      str(ex), allow_copy=True)
+                            logger.exception("Can't load existing files or commands")
+                            self.cleanup()
+                            return
             try:
                 for key, command_list in self.preloaded_commands.items():
                     logger.info(f"Attempting to apply commands from {key}")
@@ -402,7 +570,8 @@ class ModdingTools(ft.Column):
                     self.preloaded_counter.count += len(command_list)
                 await asyncio.sleep(0.001)
             except Exception as ex:
-                await self.app.show_alert(tr("unable_to_apply_commands"), str(ex),
+                await self.app.show_alert(tr("unable_to_apply_commands",
+                                             target=self.source_path_field.value), str(ex),
                                           allow_copy=True)
                 logger.exception("Unable to apply commands to source tree")
                 self.cleanup()
@@ -477,7 +646,7 @@ class ModdingTools(ft.Column):
         try:
             for cmd in commands_generator:
                 total_processed += 1
-                if (time.perf_counter() - last_upd_nodes) > 0.05:
+                if (time.perf_counter() - last_upd_nodes) > self.node_counter.update_timeout:
                     self.node_counter.count = total_processed
                     await asyncio.sleep(0.001)
                     last_upd_nodes = time.perf_counter()
@@ -488,7 +657,7 @@ class ModdingTools(ft.Column):
                 self.commands.append(cmd)
                 self.commands_view.controls.append(CommandCard(self, cmd))
 
-                if (time.perf_counter() - last_upd_cmds) > 0.05:
+                if (time.perf_counter() - last_upd_cmds) > self.node_counter.update_timeout:
                     self.command_counter.count = len(self.commands)
                     self.commands_view.update()
                     await asyncio.sleep(0.001)
@@ -515,23 +684,28 @@ class ModdingTools(ft.Column):
         self.toggle_output_btns(enable=True)
 
 class CommandPreview(ft.Container):
-    def __init__(self, modding_tools: "ModdingTools") -> None:
+    def __init__(self, modding_tools: "MergeTool") -> None:
         super().__init__()
         self.modding_tools = modding_tools
         self.expand = True
 
     def before_update(self) -> None:
         if self.modding_tools.current_command_card is None:
-            self.content = ft.Row([ft.Column([
-                    ft.Placeholder(color=ft.Colors.SECONDARY_CONTAINER)],
-                expand=True)],
-            expand=True)
+            self.content = ft.Column([
+                    ft.Row([
+                        ft.Container(
+                           ft.Placeholder(color=ft.Colors.SECONDARY_CONTAINER),
+                            margin=ft.margin.symmetric(horizontal=10, vertical=3),
+                            clip_behavior=ft.ClipBehavior.NONE)],
+                        wrap=True)],
+                    scroll=ft.ScrollMode.AUTO)
             return
 
         cmd = self.modding_tools.current_command_card.command
-        serizalized_cmd = xml_diff.Differ.serialize_command(cmd, keep_attrs=["_DuplicateCount"])
+        serizalized_cmd = xml_diff.Differ.serialize_command(cmd) #, keep_attrs=["_DuplicateCount"])
         objectify.deannotate(serizalized_cmd, cleanup_namespaces=True)
 
+        etree.indent(serizalized_cmd, space="    ")
         cmd_xml_string = \
             parse_ops.beautify_machina_xml(
                 etree.tostring(
@@ -544,12 +718,8 @@ class CommandPreview(ft.Container):
                 ft.Column([
                     ft.Row([
                         ft.Container(
-                           ft.Markdown(
+                           cw.MarkdownWithCode(
                                 f"```xml\n{cmd_xml_string}\n```",
-                                extension_set=ft.MarkdownExtensionSet.GITHUB_WEB,
-                                code_style_sheet=ft.MarkdownStyleSheet(
-                                code_text_style=ft.TextStyle(color=ft.Colors.WHITE70)),
-                                code_theme=ft.MarkdownCodeTheme.ATOM_ONE_DARK,
                                 selectable=True),
                             margin=ft.margin.only(right=15),
                             border_radius=10, clip_behavior=ft.ClipBehavior.ANTI_ALIAS)],
@@ -557,7 +727,7 @@ class CommandPreview(ft.Container):
                     scroll=ft.ScrollMode.ALWAYS)
 
 class NodePreview(ft.Container):
-    def __init__(self, modding_tools: "ModdingTools",
+    def __init__(self, modding_tools: "MergeTool",
                  node: objectify.ObjectifiedElement | None = None,
                  file_path: Path | None = None,
                   **kwargs) -> None:
@@ -567,9 +737,6 @@ class NodePreview(ft.Container):
         self.node = node
         self.count = 1
         self.expand = True
-        self.content = ft.Container(
-            ft.Placeholder(color=ft.Colors.SECONDARY_CONTAINER),
-            padding=10)
         self.edit_btn = ft.IconButton(
             icon=ft.Icons.EDIT_NOTE,
             on_click=self.open_path,
@@ -586,13 +753,20 @@ class NodePreview(ft.Container):
             file_ops.open_file_in_editor(self.file_path, line=int(self.node.sourceline),
                                          editor=self.modding_tools.app.config.code_editor)
 
+    def reset(self) -> None:
+        self.content = ft.Container(
+            ft.Placeholder(color=ft.Colors.SECONDARY_CONTAINER),
+            padding=ft.padding.symmetric(horizontal=10, vertical=3))
+
     def before_update(self) -> None:
         if not self.modding_tools.current_command_card:
+            self.reset()
             return
 
         self.edit_btn.disabled = self.node is None
 
         if self.node is not None:
+            etree.indent(self.node, space="    ")
             node_string = parse_ops.beautify_machina_xml(
                 etree.tostring(
                     self.node,
@@ -604,12 +778,8 @@ class NodePreview(ft.Container):
                     ft.Column([
                         ft.Row([
                             ft.Container(
-                                ft.Markdown(
+                                cw.MarkdownWithCode(
                                     f"```html\n{node_string}\n```",
-                                    extension_set=ft.MarkdownExtensionSet.GITHUB_WEB,
-                                    code_style_sheet=ft.MarkdownStyleSheet(
-                                        code_text_style=ft.TextStyle(color=ft.Colors.WHITE70)),
-                                    code_theme=ft.MarkdownCodeTheme.ATOM_ONE_DARK_REASONABLE,
                                     selectable=True),
                                 margin=ft.margin.only(right=15),
                                 border_radius=10,
@@ -635,8 +805,9 @@ class NodePreview(ft.Container):
                 ft.Column([
                     ft.Row([
                         ft.Container(
-                            ft.Text("Node not found",
+                            ft.Text(tr("node_not_found"),
                                 color=ft.Colors.RED,
+                                font_family="Fira Code",
                                 weight=ft.FontWeight.W_500,
                                 text_align=ft.TextAlign.CENTER,
                                 # expand=True
@@ -649,6 +820,7 @@ class NodePreview(ft.Container):
 class CircleCounter(ft.Stack):
     def __init__(self, default_bg_color: ft.Colors = ft.Colors.RED, **kwargs) -> None:
         super().__init__(**kwargs)
+        self.update_timeout = 0.05
         self.alignment = ft.alignment.center
         self._count = 0
         self.default_bg_color = default_bg_color
@@ -703,7 +875,7 @@ class CircleCounter(ft.Stack):
             self.ring.visible = False
 
 class CommandCard(ft.Card):
-    def __init__(self, modding_tools: "ModdingTools", command: Command, **kwargs):
+    def __init__(self, modding_tools: "MergeTool", command: Command, **kwargs) -> None:
         super().__init__(**kwargs, elevation=7)
         self.modding_tools = modding_tools
         self.command = command
@@ -711,7 +883,7 @@ class CommandCard(ft.Card):
                               ActionType.REPLACE,
                               ActionType.ADD_OR_REPLACE]:
             type_color = ft.Colors.GREEN
-        elif command.action == ActionType.MODIFY:
+        elif command.action in [ActionType.MODIFY, ActionType.MODIFY_OR_FAIL]:
             type_color = ft.Colors.ORANGE
         else:
             type_color = ft.Colors.RED
@@ -719,12 +891,16 @@ class CommandCard(ft.Card):
         self.surface_tint_color = type_color
 
         self.checkbox = ft.Checkbox(scale=0.9)
-        self.content = ft.Container(ft.Column([
-            ft.Row([self.checkbox,
-                    ft.Text(str(command.action.value), weight=ft.FontWeight.W_700, color=type_color)]),
-            ft.Text(command.selector),
-            ft.Text(f"Parent: {command.parent_path or 'root'}", opacity=0.8),
-            ]),
-            padding=ft.padding.symmetric(5, 10), on_click=self.modding_tools.click_command)
+        self.content = ft.GestureDetector(
+            ft.Container(ft.Column([
+                ft.Row([self.checkbox,
+                        ft.Text(str(command.action.value), weight=ft.FontWeight.W_700, color=type_color)]),
+                ft.Text(command.selector),
+                ft.Text(f"Parent: {command.parent_path or 'root'}", opacity=0.8),
+                ]),
+                padding=ft.padding.symmetric(5, 10)),
+            on_tap=self.modding_tools.click_command,
+            on_secondary_tap=self.modding_tools.secondary_click_command,
+            mouse_cursor=ft.MouseCursor.CLICK)
 
 
