@@ -1,3 +1,8 @@
+from pathlib import Path
+
+from commod.localisation.service import tr
+
+
 class DistributionNotFoundError(Exception):
     def __init__(self, path: str, message: str = "Invalid distibution path") -> None:
         self.path = path
@@ -78,5 +83,29 @@ class InvalidExistingManifestError(Exception):
     def __str__(self) -> str:
         return f"Manifest is invalid: '{self.manifest_path}'"
 
-class ModFileInstallationError(Exception):
+class ModMissingFileInstallationError(Exception):
+    def __init__(self, problematic_file: Path | str) -> None:
+        self.problematic_file = problematic_file
+        super().__init__(self.problematic_file)
+
+    def __str__(self) -> str:
+        return (tr("missing_target_file_for_merge_command",
+                  target=str(Path("data") / self.problematic_file))
+                + "\n\n" + tr("mod_is_incompatible_with_current_game")
+                + "\n\n" + tr("debug_info_is_available"))
+
+class ModInvalidMergeInstallationError(Exception):
+    def __init__(self, problematic_file: Path | str, command_str: str | None = None) -> None:
+        self.problematic_file = problematic_file
+        self.command_str = command_str
+        super().__init__(self.problematic_file)
+
+    def __str__(self) -> str:
+        return (tr("unable_to_apply_commands",
+                  target=str(self.problematic_file))
+                + (f"\n{self.command_str}" if self.command_str else "")
+                + "\n\n" + tr("mod_is_incompatible_with_current_game")
+                + "\n\n" + tr("debug_info_is_available"))
+
+class ModFilePackagingError(Exception):
     pass
