@@ -1,7 +1,6 @@
 import argparse
 import html
-import xml.etree.ElementTree as ET
-
+import re
 from collections.abc import Iterable
 from functools import cache
 from pathlib import Path
@@ -115,127 +114,135 @@ def find_element(
                   f"'{xml_node.tag}' in '{xml_node.base}'")
         return None
 
-def find_node(
-        xml_node: ET.Element | None,
-        child_name: str, do_not_warn: bool = False) -> ET.Element | None:
-    """Get the first child from ElementTree by the name."""
-    if not isinstance(xml_node, ET.Element):
-        if not do_not_warn:
-            print(f"Can't find child node, as parent element is invalid: '{xml_node}'")
-        return None
-    child = xml_node.find(child_name)
-    if child is None and not do_not_warn:
-        print(f"No child with name '{child_name}' for xml node '{xml_node.tag}'")
-    return child
-
-def find_nodes(
-        xml_node: ET.Element | None,
-        children_name: str, do_not_warn: bool = False) -> list[ET.Element]:
-    """Get all matching children from ElementTree by their name."""
-    if not isinstance(xml_node, ET.Element):
-        if not do_not_warn:
-            print(f"Can't find child node, as parent element is invalid: '{xml_node}'")
-        return []
-    childs = xml_node.findall(children_name)
-    if not childs and not do_not_warn:
-        print(f"No children with name '{children_name}' for xml node '{xml_node.tag}'")
-    return childs
-
-def get_attrib(
-        xml_node: ET.Element | None,
-        attrib_name: str, do_not_warn: bool = False) -> str | None:
-    if not isinstance(xml_node, ET.Element):
-        if not do_not_warn:
-            print(f"Can't get attribute '{attrib_name}', node is invalid: '{xml_node}'")
-        return None
-    attrib = xml_node.get(attrib_name)
-    if attrib is None and not do_not_warn:
-        print(f"No attribute with name '{attrib_name}' for xml node '{xml_node.tag}'")
-    return attrib
 
 def beautify_machina_xml(xml_string: bytes) -> bytes:
-    """Format and beautify xml string in the style similar to original Ex Machina dynamicscene.xml files."""
-    beautified_string = b""
-    previous_line_indent = -1
+    """Format and beautify XML string in the style similar to original Ex Machina dynamicscene.xml files."""
+    never_split_tags = {b"event", b"Point", b"Wheel"}
+    xml_declaration_prefix = b"<?xml"
+    xml_delcaration_suffix = b"?>"
+    script_start_tag = b"<script>"
+    script_end_tag = b"</script>"
+    indent_symbol = b"    "
 
-    # As first line of xml file is XML Declaration, we want to exclude it
-    # from Beautifier to get rid of checks for every line down the line
+    beautified_parts = []
+    previous_line_indent = -1
     inside_plaintext_block = False
     plaintext_tag_indent = 0
-
-    never_split_tags = [b"event", b"Point", b"Wheel"] # b"Folder"
+    plaintext_block_offset = None
 
     for raw_line in xml_string.splitlines():
-        line = raw_line
+        line_stripped = raw_line.lstrip()
 
-        line_stripped = line.lstrip()
+        if (line_stripped.startswith(xml_declaration_prefix)
+           and line_stripped.endswith(xml_delcaration_suffix)):
+            beautified_parts.append(raw_line + b"\n\n")
+            continue
 
         # calculating indent level of parent line to indent attributes
-        # lxml use spaces for indents, game use tabs, so indents maps 2:1
-        line_indent = (len(line) - len(line_stripped)) // 2
+        # We use 4 spaces for indents
+        line_indent = (len(raw_line.expandtabs(4)) - len(line_stripped)) // 4
 
-        if line_stripped.startswith(b"</script>"):
-            inside_plaintext_block = False
-            line_indent = plaintext_tag_indent
-            previous_line_indent = 0
-            plaintext_tag_indent = 0
-
-        if line_stripped.startswith(b"<?xml") and line_stripped.endswith(b"?>"):
-            beautified_string += line + b"\n\n"
-            continue
-
+        # Content in plain text blocks (scripts in triggers) will not be treated as xml,
+        # but we still need to indent it correctly.
+        # We will make sure that script is always indented once, not more or less
         if inside_plaintext_block:
-            beautified_string += line + b"\n"
-            continue
+            if line_stripped.startswith(script_end_tag):
+                inside_plaintext_block = False
+                line_indent = plaintext_tag_indent
+                previous_line_indent = 0
+                plaintext_tag_indent = 0
+                plaintext_block_offset = None
+            else:
+                if plaintext_block_offset is None:
+                    plaintext_block_offset = previous_line_indent - line_indent + 1
 
+                detabbed_line = raw_line.expandtabs(4) + b"\n"
+                if detabbed_line.isspace():
+                    beautified_parts.append(b"\n")
+                    continue
 
-        # beautified_string += line_indent * b"\t" + line + b"\n"
+                if plaintext_block_offset < 0:
+                    detabbed_line = detabbed_line[len(indent_symbol) * abs(plaintext_block_offset):]
+                else:
+                    detabbed_line = indent_symbol * plaintext_block_offset + detabbed_line
+                beautified_parts.append(detabbed_line)
 
-        if any(line_stripped.startswith(b"<"+prefix) for prefix in never_split_tags):
-            line = line_indent * b"\t" + line_stripped + b"\n"
+                continue
+
+        # Format line based on tag type \ node being comment
+        formatted_line = None
+
+        tag_prefix = _extract_tag_prefix(line_stripped)
+        if line_stripped.startswith(b"<!--") or tag_prefix in never_split_tags:
+            formatted_line = line_indent * indent_symbol + line_stripped + b"\n"
         else:
-            # manually tabulating lines according to saved indent level
-            line = _split_tag_on_attributes(line_stripped, line_indent)
-            line = line_indent * b"\t" + line + b"\n"
+            formatted_line = _split_tag_on_attributes(line_stripped, line_indent)
+            formatted_line = line_indent * indent_symbol + formatted_line + b"\n"
 
-        # in EM xmls every first and only first tag of its tree level is
-        # separated by a new line
+        # Add blank line before tags at the same level (first tag of its tree level)
         if line_indent == previous_line_indent:
-            line = b"\n" + line
+            formatted_line = b"\n" + formatted_line
 
-        # we need to know indentation of previous tag to decide if tag is
-        # first for its tree level, as described above
+        beautified_parts.append(formatted_line)
         previous_line_indent = line_indent
 
-        beautified_string += line
-
-        if line_stripped.startswith(b"<script>"):
+        if line_stripped.startswith(script_start_tag):
             inside_plaintext_block = True
             plaintext_tag_indent = line_indent
 
-    return beautified_string
+    return b"".join(beautified_parts)
 
+def _extract_tag_prefix(tag_line: bytes) -> bytes:
+    """
+    Extract the tag name from an XML tag line.
+
+    Example: b"<Point x=10>" -> b"Point".
+    """
+    if not tag_line.startswith(b"<"):
+        return b""
+
+    # Find end of tag name (space or closing bracket)
+    end_pos = tag_line.find(b" ")
+    if end_pos == -1:
+        end_pos = tag_line.find(b">")
+    if end_pos == -1:
+        return b""
+
+    # Extract just the tag name
+    tag_name = tag_line[1:end_pos]
+
+    # Remove any namespace prefix if present
+    colon_pos = tag_name.find(b":")
+    if colon_pos != -1:
+        tag_name = tag_name[colon_pos+1:]
+
+    return tag_name
 
 def _split_tag_on_attributes(xml_line: bytes, line_indent: int) -> bytes:
-    white_space_index = xml_line.find(b" ")
-    quotmark_index = xml_line.find(b'"')
-
-    # true when no tag attribute contained in string
-    if white_space_index == -1 or quotmark_index == -1:
+    # Skip processing for comments
+    indent_symbol = b"    "
+    if xml_line.startswith(b"<!--") and b"-->" in xml_line:
         return xml_line
 
-    if white_space_index < quotmark_index:
-        # next tag attribute found, now indent found attribute and
-        # recursively start work on a next line part
-        return (xml_line[:white_space_index] + b"\n" + b"\t" * (line_indent + 1)
-                + _split_tag_on_attributes(xml_line[white_space_index + 1:],
-                                           line_indent))
+    if b" " not in xml_line or b'"' not in xml_line:
+        return xml_line
 
-    # searching where attribute values ends and new attribute starts
-    second_quotmark_index = xml_line.find(b'"', quotmark_index + 1) + 1
-    return (xml_line[:second_quotmark_index]
-            + _split_tag_on_attributes(xml_line[second_quotmark_index:],
-                                       line_indent))
+    # Find tag name first
+    tag_match = re.match(rb"[^\s]+", xml_line)
+    if not tag_match:
+        return xml_line
+
+    tag_name = tag_match.group(0)
+    remainder = xml_line[len(tag_name):]
+
+    # Pattern to match: space followed by attr="value"
+    # Using positive lookahead to find spaces that are followed by an attribute
+    attr_pattern = rb'(?=\s+[^\s="]+="[^"]*")'
+
+    parts = map(bytes.strip, re.split(attr_pattern, remainder))
+
+    indent = b"\n" + indent_symbol * (line_indent + 1)
+    return tag_name + indent.join(parts)
 
 
 def xml_to_objfy(full_path: str | Path) -> objectify.ObjectifiedElement:
@@ -243,17 +250,13 @@ def xml_to_objfy(full_path: str | Path) -> objectify.ObjectifiedElement:
         byte_string = fh.read()
         try:
             encoding = "utf-8"
-            file_data = byte_string.decode(encoding)
+            byte_string.decode(encoding)
         except UnicodeDecodeError:
             encoding = data.ENCODING
-            file_data = byte_string.decode(data.ENCODING)
+            byte_string.decode(data.ENCODING)
 
         parser_recovery = objectify.makeparser(recover=True, encoding=encoding, collect_ids=False)
         objectify.enable_recursive_str(True)
-        # objectify.parse(f, parser_recovery)
-        objfy = objectify.fromstring(byte_string, parser_recovery)
-    return objfy
 
-def xml_to_etree(full_path: str | Path) -> ET.Element:
-    tree = ET.parse(full_path)
-    return tree.getroot()
+        # objectify.parse(f, parser_recovery)
+        return objectify.fromstring(byte_string, parser_recovery)
