@@ -1,47 +1,20 @@
-from collections.abc import Iterable
-from dataclasses import dataclass
-from enum import Enum
+import contextlib
+from copy import deepcopy
 from pathlib import Path
 
 from lxml import etree, objectify
 
 from commod.helpers import file_ops, parse_ops
+from commod.localisation.service import tr
+from commod.tools.xml_diff import Differ
+from commod.tools.xml_helpers import ActionType, AmbiguousMergeCommandError, Command, InvalidMergeCommandError
 
-
-class InvalidMergeCommandError(Exception):
-    ...
-
-class AmbiguousMergeCommandError(Exception):
-    ...
-
-class ActionType(Enum):
-    ADD = "Add"
-    REPLACE = "Replace"
-    ADD_OR_REPLACE = "AddOrReplace"
-    MODIFY = "Modify"
-    REMOVE = "Remove"
-
-    @classmethod
-    def list_values(cls) -> list[str]:
-        return [c.value for c in cls]
+COMMAND_LIMIT = 100000
 
 def reformat_xml(input_path: Path, output_path: Path) -> None:
     input_tree = parse_ops.xml_to_objfy(input_path)
     file_ops.write_xml_to_file(input_tree, output_path, machina_beautify=True, use_utf=False)
 
-@dataclass
-class Command:
-    action: ActionType
-    parent_path: str
-    selector: str
-    tag: str
-    node_attrs: dict
-    selector_keys: Iterable[str]
-    children_nodes: list[objectify.ObjectifiedElement] | None
-    merge_author: str = ""
-    desired_count: int = 1
-    source_node: objectify.ObjectifiedElement | None = None
-    modded_node: objectify.ObjectifiedElement | None = None
 
 def traverse_path(tree: objectify.ObjectifiedElement, full_path: str) -> objectify.ObjectifiedElement:
     element = tree
@@ -83,7 +56,11 @@ def parse_commands(xml_node: objectify.ObjectifiedElement,
     if action_parsed and action_parsed not in ActionType.list_values():
         raise InvalidMergeCommandError(f"Invalid action type: '{action_parsed}'")
 
-    action = ActionType(action_parsed)
+    try:
+        action = ActionType(action_parsed)
+    except ValueError as ex:
+        raise InvalidMergeCommandError(
+            f"Can't load command from node with tag '{xml_node.tag}' as it's missing _Action") from ex
 
     if selector_keys_parsed:
         if selector_keys_parsed == "*":
@@ -98,14 +75,19 @@ def parse_commands(xml_node: objectify.ObjectifiedElement,
     if selector_keys and selector:
         raise InvalidMergeCommandError("Defined both _Selector and _SelectorKeys, need only one")
 
-    if action in [ActionType.REPLACE, ActionType.ADD_OR_REPLACE,
-                  ActionType.MODIFY, ActionType.REMOVE] and not (selector_keys or selector):
+    # action in [ActionType.REPLACE, ActionType.ADD_OR_REPLACE,
+            #   ActionType.MODIFY, ActionType.REMOVE] and
+    if not (selector_keys or selector):
         raise InvalidMergeCommandError(f"'{action}' actions must define _Selector or _SelectorKeys attribute")
 
     if isinstance(desired_count, str) and not desired_count.isnumeric():
         raise InvalidMergeCommandError(f"_DesiredCount must be numeric instead of '{desired_count}'")
 
+
     desired_count = int(desired_count)
+
+    if desired_count > 1 and action in [ActionType.MODIFY, ActionType.MODIFY_OR_FAIL]:
+        raise InvalidMergeCommandError("Modify commands can't specify ActionType > 1")
 
     max_desired_count = 50
     if desired_count > max_desired_count:
@@ -117,7 +99,9 @@ def parse_commands(xml_node: objectify.ObjectifiedElement,
         raise InvalidMergeCommandError(f"'{action}' actions can't have children nodes")
 
     if action:
-        # maybe need to implicitly add/merge parent even when _Action is not specified
+        # TODO: maybe need to support nested commands
+        if len(commands) > COMMAND_LIMIT:
+            raise ValueError("Too many commands created")
         commands.append(
             Command(
                 action=action,
@@ -144,7 +128,7 @@ def parse_commands(xml_node: objectify.ObjectifiedElement,
 
             parse_commands(child, commands, new_path, merge_author)
     # we directy add all children nodes for Add actions, without processing them separately
-    # if action == ActionType.Add:
+    # if action == ActionType.ADD:
         # return commands
 
     return commands
@@ -153,7 +137,7 @@ def apply_command(tree: objectify.ObjectifiedElement, command: Command) -> objec
     # if not command.parent_path.startswith(f"//{tree.tag}"):
         # raise InvalidMergeCommandError
 
-    elm = traverse_path(tree, command.parent_path) if command.parent_path else tree
+    base_element = traverse_path(tree, command.parent_path) if command.parent_path else tree
 
     if command.selector_keys:
         selector = command.tag + "".join(f'[@{key}="{command.node_attrs.get(key)}"]'
@@ -161,7 +145,10 @@ def apply_command(tree: objectify.ObjectifiedElement, command: Command) -> objec
     else:
         selector = command.selector
 
-    elements = elm.xpath(selector)
+    try:
+        elements = base_element.xpath(selector)
+    except Exception as ex:
+        raise InvalidMergeCommandError(f"Command with invalid selector: {command}") from ex
 
     if command.action == ActionType.ADD_OR_REPLACE:
         if not elements:
@@ -170,46 +157,67 @@ def apply_command(tree: objectify.ObjectifiedElement, command: Command) -> objec
             command.action = ActionType.REPLACE
         apply_command(tree, command)
     elif command.action == ActionType.ADD:
+        new_elm = objectify.Element(command.tag)
+        for attr_key, attr_val in command.node_attrs.items():
+            if not attr_key.startswith("_"):
+                new_elm.set(attr_key, attr_val)
+        new_elm.set("_MergeAuthor", command.merge_author)
+        if command.children_nodes:
+            new_elm.extend(command.children_nodes)
+        objectify.deannotate(new_elm, cleanup_namespaces=True)
+
         if elements:
-            raise InvalidMergeCommandError(
-                f"Can't Add node, matching element already exist for selector '{selector}'")
+            if command.desired_count == 1:
+                if Differ.are_equivalent_nodes(elements[0], new_elm):
+                    return tree
 
-        for _ in range(command.desired_count):
-            new_elm = objectify.Element(command.tag)
-            for attr_key, attr_val in command.node_attrs.items():
-                if not attr_key.startswith("_"):
-                    new_elm.set(attr_key, attr_val)
-            new_elm.set("_MergeAuthor", command.merge_author)
+                err_msg = tr("cant_add_node", selector=selector)
 
-            if command.children_nodes:
-                new_elm.extend(command.children_nodes)
+                node_last_author = ""
+                with contextlib.suppress(Exception):
+                    node_last_author = elements[0].get("_MergeAuthor")
+                if node_last_author:
+                    err_msg += tr("source_node_last_modifier", node_last_author=node_last_author)
+                else:
+                    err_msg += tr("source_node_created_with_no_merge")
+                raise InvalidMergeCommandError(err_msg)
 
-            elm.append(new_elm)
+            if command.desired_count < len(elements):
+                while command.desired_count != len(elements):
+                    existing_elem = elements.pop()
+                    base_element.remove(existing_elem)
+                return tree
+
+        for _ in range(command.desired_count - len(elements)):
+            base_element.append(deepcopy(new_elm))
+
     elif command.action == ActionType.REPLACE:
+        new_elm = objectify.Element(command.tag)
+        for attr_key, attr_val in command.node_attrs.items():
+            if not attr_key.startswith("_"):
+                new_elm.set(attr_key, attr_val)
+        new_elm.set("_MergeAuthor", command.merge_author)
+        if command.children_nodes:
+            new_elm.extend(command.children_nodes)
+        objectify.deannotate(new_elm, cleanup_namespaces=True)
+
         if not elements:
             raise InvalidMergeCommandError(
                 f"Can't Replace node, nothing matches selector '{selector}'")
-        if len(elements) > 1:
-            raise InvalidMergeCommandError(
-                f"Can't Replace node, multiple matches found for selector '{selector}'")
-        element = elements[0]
-        element.attrib.clear()
 
-        for child in element.getchildren():
-            element.remove(child)
+        while elements:
+            existing_elem = elements.pop()
+            base_element.remove(existing_elem)
 
-        for attr_key, attr_val in command.node_attrs.items():
-            if not attr_key.startswith("_"):
-                element.set(attr_key, attr_val)
-        element.set("_MergeAuthor", command.merge_author)
+        for _ in range(command.desired_count):
+            base_element.append(deepcopy(new_elm))
 
-        if command.children_nodes:
-            element.extend(command.children_nodes)
-
-    elif command.action == ActionType.MODIFY:
+    elif command.action in [ActionType.MODIFY, ActionType.MODIFY_OR_FAIL]:
         if not elements:
-            raise InvalidMergeCommandError(
-                f"Can't Modify node, nothing matches selector '{selector}'")
+            if command.action == ActionType.MODIFY_OR_FAIL:
+                raise InvalidMergeCommandError(
+                    f"Can't Modify node, nothing matches selector '{selector}'")
+            return tree
         for elem in elements:
             for attr_key, attr_val in command.node_attrs.items():
                 if not attr_key.startswith("_") and attr_key not in command.selector_keys:
@@ -217,17 +225,27 @@ def apply_command(tree: objectify.ObjectifiedElement, command: Command) -> objec
             elem.set("_MergeAuthor", command.merge_author)
     elif command.action == ActionType.REMOVE:
         if not elements:
-            print(f"Remove command is useless, had no matching nodes for selector '{selector}' anyway")
+            if command.action == ActionType.REMOVE_OR_FAIL:
+                raise InvalidMergeCommandError(
+                    f"Can't Remove node, nothing matches selector '{selector}'")
+            # print(f"Remove command is useless, had no matching nodes for selector '{selector}' anyway")
             return tree
-        element = elements[0]
-        elm.remove(element)
+        while elements:
+            element = elements.pop()
+            base_element.remove(element)
     else:
         raise InvalidMergeCommandError(f"Unknown merge command: {command.action}")
     return tree
 
 def apply_commands(base_tree: objectify.ObjectifiedElement, commands: list[Command]) -> objectify.ObjectifiedElement:
     for command in commands:
-        apply_command(base_tree, command)
+        try:
+            apply_command(base_tree, command)
+        except InvalidMergeCommandError as ex:
+            raise InvalidMergeCommandError((f"{ex.error_desc}\n\n" if ex.error_desc else "")  # noqa: B904
+                                           + f"{command!s}")
+        except Exception as ex:
+            raise InvalidMergeCommandError(str(command)) from ex
     objectify.deannotate(base_tree, cleanup_namespaces=True)
     return base_tree
 
@@ -254,3 +272,8 @@ def combine_two_xmls(base_path: Path, commands_file: Path,
     updated_tree = apply_commands(base_tree, commands)
     file_ops.write_xml_to_file(updated_tree, output_path, machina_beautify=True)
     print(f"Combined xmls, output: {output_path}")
+
+def update_file_with_commands(source_path: Path, commands: list[Command]) -> None:
+    source_tree = parse_ops.xml_to_objfy(source_path)
+    updated_tree = apply_commands(source_tree, commands)
+    file_ops.write_xml_to_file(updated_tree, source_path, machina_beautify=True)
